Cookbook: remove Misc category
