+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/General/GridRecipes.cs
@@ -1,6 +1,4 @@
-﻿using System.ComponentModel;
-
-namespace ScottPlotCookbook.Recipes.Axis;
+﻿namespace ScottPlotCookbook.Recipes.Axis;
 
 public class CustomizingGrids : ICategory
 {
