+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/Miscellaneous/Internationalization.cs
@@ -2,7 +2,7 @@
 
 public class Internationalization : ICategory
 {
-    public Chapter Chapter => Chapter.Misc;
+    public Chapter Chapter => Chapter.General;
     public string CategoryName => "Internationalization";
     public string CategoryDescription => "Using ScottPlot across cultures with different text and numeric requirements.";
